Moving most of RouteHandler component to a RouteHandler mixin
