+++ .gitignore
@@ -1,1 +1,2 @@
 examples/**/*-bundle.js
+node_modules
