+++ modules/components/RouteHandler.js
@@ -1,4 +1,5 @@
 var React = require('react');
+var RouteHandlerMixin = require('../mixins/RouteHandler');
 
 /**
  * A <RouteHandler> component renders the active child route handler
@@ -8,49 +9,16 @@ var RouteHandler = React.createClass({
 
   displayName: 'RouteHandler',
 
+  mixins: [RouteHandlerMixin],
+
   getDefaultProps: function () {
     return {
       ref: '__routeHandler__'
     };
   },
 
-  contextTypes: {
-    getRouteAtDepth: React.PropTypes.func.isRequired,
-    getRouteComponents: React.PropTypes.func.isRequired,
-    routeHandlers: React.PropTypes.array.isRequired
-  },
-
-  childContextTypes: {
-    routeHandlers: React.PropTypes.array.isRequired
-  },
-
-  getChildContext: function () {
-    return {
-      routeHandlers: this.context.routeHandlers.concat([ this ])
-    };
-  },
-
-  getRouteDepth: function () {
-    return this.context.routeHandlers.length - 1;
-  },
-
-  componentDidMount: function () {
-    this._updateRouteComponent();
-  },
-
-  componentDidUpdate: function () {
-    this._updateRouteComponent();
-  },
-
-  _updateRouteComponent: function () {
-    var depth = this.getRouteDepth();
-    var components = this.context.getRouteComponents();
-    components[depth] = this.refs[this.props.ref];
-  },
-
   render: function () {
-    var route = this.context.getRouteAtDepth(this.getRouteDepth());
-    return route ? React.createElement(route.handler, this.props) : null;
+    return this.getRouteHandler();
   }
 
 });
